Release version 23.03.2 (#23)
* Version 22.11.1: Fix Bug/12 summary results (#13)

* fix export of interval data

* add colnames to export interval data, shift UI to load model above model tabs, fix error when export cred intervals plot

* version up, add news.md

* use indVar input for default model name

* show always most recent result by default

* update news

* update news.md

* fix reactive behavior in the tab credibility intervals over time, update remote test data

* clean up after model upload, update readme

* update news.md

* fix update of min/max time in Time point estimates

* update min/max time in estimates for user defined interval

* update news.md

* Update README.md

* apply import module (#16)

* apply import module for Renewal Rates

* use import module for Measurements

* use import module for residence time

* use import module in measurements tab,
fixed sidebars with auto-scroll

* update UI, fix warning

* version up, update news.md

* update namespace

* Update README.md

* Update README.md

* Update README.md

* more try catch staements (#17)

* add remote package for gpt3

* trigger build

* Feat/merge renew unc (#21)

* prepare merging changes from main into beta

* merge fix import

* apply importData module for renewal rates uncertainty

---------


* load models directly from github (#22)

* Update README.md

* Version 23.04.1: Fix logic if renewal rates uncertainty is unchecked (#25)

* if no uncertainty use zero uncertainty matrix, remove duplicated code, extract clean and split of data, add test for cleanAndSplitData

* extract update of uncertainty matrix, extract clean and split of data and uncertainty matrix per individuum, update tests

* add error message

* update news.md

* add missing export

* fix warning

* add missing export

* update namespace

---------

+++ R/02-downAndUploadModel.R
@@ -17,7 +17,7 @@ downloadModelUI <- function(id, label) {
                 choices = NULL,
                 options = list(`actions-box` = TRUE),
                 multiple = T),
-    checkboxInput(ns("onlyInputs"), "Store only data and model options"),
+    checkboxInput(ns("onlyInputs"), "Store only input data and model options"),
     textAreaInput(ns("notes"), "Notes"),
     HTML("<br>"),
     downloadButton(ns("downloadModel"), "Download"),
@@ -95,18 +95,10 @@ uploadModelUI <- function(id, label) {
   ns <- NS(id)
   
   tagList(
-    HTML("<br>"),
     tags$h5(label),
-    fileInput(ns("uploadModel"), label = NULL),
-    selectInput(
-      ns("remoteModel"),
-      label = "Select remote model",
-      choices = dir(file.path(settings$pathToSavedModels)) %>%
-        sub(pattern = '\\.zip$', replacement = ''),
-      selected = NULL
-    ),
-    actionButton(ns("loadRemoteModel"), "Load Remote Model")#,
-    #helpText("Remote models are only available on on https://isomemoapp.com")
+    fileInput(ns("uploadModel"), label = "Load local model"),
+    remoteModelsUI(ns("remoteModels")),
+    tags$br(), tags$br(), tags$br()
   )
 }
 
@@ -135,8 +127,15 @@ uploadModel <- function(input, output, session, savedModels, uploadedNotes, fit,
     pathToModel(input$uploadModel$datapath)
   })
   
-  observeEvent(input$loadRemoteModel, {
-    pathToModel(file.path(settings$pathToSavedModels, paste0(input$remoteModel, ".zip")))
+  pathToRemote <- remoteModelsServer("remoteModels",
+                                     githubRepo = "osteo-bior",
+                                     rPackageName = "OsteoBioR",
+                                     rPackageVersion = "OsteoBioR" %>%
+                                       packageVersion() %>%
+                                       as.character())
+  
+  observeEvent(pathToRemote(), {
+    pathToModel(pathToRemote())
   })
   
   observeEvent(pathToModel(), {
@@ -154,10 +153,10 @@ uploadModel <- function(input, output, session, savedModels, uploadedNotes, fit,
       return()
     }
     
-    file.remove("model.rds")
-    file.remove("README.txt")
-    file.remove("help.html")
-    
+    if (file.exists("model.rds")) file.remove("model.rds")
+    if (file.exists("README.txt")) file.remove("README.txt")
+    if (file.exists("help.html")) file.remove("help.html")
+
     if (!exists("modelImport")) {
       shinyjs::alert("File format not valid. Model object not found.")
       return()
